+++ Adj/Program/Primitive/Generation.hs
@@ -3,7 +3,7 @@
 module Adj.Program.Primitive.Generation where
 
 import Adj.Auxiliary (Casted, Casting ((=-), (-=)), type (=!?=), FG (FG), FFGH (FFGH), type (=!!??=), Structural (Structural))
-import Adj.Algebra.Category (Semigroupoid ((.)), Category ((.:), (...:), (....:), (.....:)), Functor (map), Covariant, Component (component), Identity (Identity), Day (Day), type (-->), Flat (Flat), (-|->), (-||->))
+import Adj.Algebra.Category (Semigroupoid ((.)), Category ((.:), (...:), (....:)), Functor, Covariant, Component (component), Identity (Identity), type (-->), Flat (Flat), (-|->))
 import Adj.Algebra.Set (Setoid, (:*:) ((:*:)), (:+:) (This, That))
 
 newtype Generation p f o = Generation
@@ -24,12 +24,6 @@ pattern Construct :: o -> f (Construction f o) -> Construction f o
 pattern Construct x xs <- Generation (FFGH (Identity x :*: FG xs))
 	where Construct x xs = Generation . FFGH ...: Identity x :*: FG xs
 
--- TODO: keep it unitl defining Functor instance for FFGH
-instance Covariant Functor (->) (->) f => Functor (-->) (-->) (Construction f) where
-	map (Flat m) = Flat .: \case
-		Generation (FFGH (Identity x :*: xs)) -> 
-			Generation . FFGH ...: Identity .: m x :*: (xs -|-> m)
-
 type Instruction = Generation (:+:)
 
 pattern Instruct :: f (Instruction f o) -> Instruction f o
@@ -39,12 +33,6 @@ pattern Instruct xs <- Generation (FFGH (That (FG xs)))
 pattern Load :: o -> Instruction f o
 pattern Load x <- Generation (FFGH (This (Identity x)))
 	where Load x = Generation . FFGH . This .: Identity x
-
--- TODO: keep it unitl defining Functor instance for FFGH
-instance Covariant Functor (->) (->) f => Functor (-->) (-->) (Instruction f) where
-	map (Flat m) = Flat .: \case
-		Generation (FFGH (This x)) -> Generation . FFGH . This ....: x -|-> m
-		Generation (FFGH (That xs)) -> Generation . FFGH . That ....: xs -|-> m
 
 instance Covariant Functor (->) (->) f => Component (-->) f (Instruction f) where
 	component = Flat .: \x -> Generation . FFGH . That . FG