+++ CHANGELOG.md
@@ -225,3 +225,4 @@
 * Change `Betwixt` type family definition
 
 # 0.1.1
+* Remove `Betwixt` type famyly